UserImage compo mod and moved to compo folder
